Partly fix of errors and warnings on windows

+++ plugins/json/json-utils.c
@@ -766,7 +766,7 @@ int dlite_json_get_value(void *ptr, const json_t *item,
   case dliteFloat:
     if (!json_is_real(item)) return errx(1, "expected json real");
     switch (size) {
-    case  4: *((float32_t *)ptr)  = json_real_value(item);  break;
+    case  4: *((float32_t *)ptr)  = (float32_t)json_real_value(item);  break;
     case  8: *((float64_t *)ptr)  = json_real_value(item);  break;
 #ifdef HAVE_FLOAT80
     case 10: *((float80_t *)ptr)  = json_real_value(item);  break;
